rename validator as api (contains public exports)
